requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cerebras-cloud-sdk==1.59.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
dotenv==0.9.9
faiss-cpu==1.11.0
filetype==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.62
langsmith==0.3.45
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
mistletoe==1.4.0
nomic==3.9.0
numpy==2.3.4
orjson==3.11.4
packaging==24.2
pandas==2.3.3
pillow==12.0.0
pyarrow==22.0.0
pydantic==2.12.4
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.1
pytz==2025.2
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
telegramify-markdown==0.5.1
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
zstandard==0.23.0
